refactor(app): keep project data out of app shell props
